addressing Roee's comments
