+++ manager/controllers/app/multicluster/multi_cluster_manager.go
@@ -1,17 +1,20 @@
 package multicluster
 
 import (
-	"fmt"
 	"github.com/ibm/the-mesh-for-data/manager/apis/app/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 )
 
-type ClusterManager interface {
+type ClusterLister interface {
 	GetClusters() ([]Cluster, error)
+}
+
+type ClusterManager interface {
+	ClusterLister
 	GetBlueprint(cluster string, namespace string, name string) (*v1alpha1.Blueprint, error)
-	CreateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) (DeploymentDetails, error)
-	UpdateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) (DeploymentDetails, error)
+	CreateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) error
+	UpdateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) error
 	DeleteBlueprint(cluster string, namespace string, name string) error
 }
 
@@ -23,45 +26,6 @@ type ClusterMetadata struct {
 type Cluster struct {
 	Name     string
 	Metadata ClusterMetadata
-}
-
-type DeploymentDetails struct {
-}
-
-type DummyMultiClusterManager struct {
-	DeployedBlueprints map[string]*v1alpha1.Blueprint
-}
-
-func (m *DummyMultiClusterManager) GetClusters() ([]Cluster, error) {
-	return []Cluster{
-		{
-			Name:     "kind-kind",
-			Metadata: ClusterMetadata{},
-		},
-	}, nil
-}
-
-func (m *DummyMultiClusterManager) GetBlueprint(cluster string, namespace string, name string) (*v1alpha1.Blueprint, error) {
-	blueprint, found := m.DeployedBlueprints[cluster]
-	if found {
-		return blueprint, nil
-	}
-	return nil, fmt.Errorf("Blueprint not found")
-}
-
-func (m *DummyMultiClusterManager) CreateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) (DeploymentDetails, error) {
-	m.DeployedBlueprints[cluster] = blueprint
-	return DeploymentDetails{}, nil
-}
-
-func (m *DummyMultiClusterManager) UpdateBlueprint(cluster string, blueprint *v1alpha1.Blueprint) (DeploymentDetails, error) {
-	m.DeployedBlueprints[cluster] = blueprint
-	return DeploymentDetails{}, nil
-}
-
-func (m *DummyMultiClusterManager) DeleteBlueprint(cluster string, namespace string, name string) error {
-	delete(m.DeployedBlueprints, cluster)
-	return nil
 }
 
 // Decode json into runtime.Object, which is a pointer (such as &corev1.ConfigMapList)
